CXX-301 Backport server r2.7.3..r2.7.4 changes
