+++ src/SConscript.client
@@ -71,6 +71,12 @@ clientSourceBasic = [
     'mongo/base/status.cpp',
     'mongo/base/string_data.cpp',
     'mongo/bson/bson_validate.cpp',
+    'mongo/bson/bsonelement.cpp',
+    'mongo/bson/bsonmisc.cpp',
+    'mongo/bson/bsonobj.cpp',
+    'mongo/bson/bsonobjbuilder.cpp',
+    'mongo/bson/bsonobjiterator.cpp',
+    'mongo/bson/bsontypes.cpp',
     'mongo/bson/oid.cpp',
     'mongo/bson/optime.cpp',
     'mongo/bson/util/bson_extract.cpp',
@@ -98,7 +104,6 @@ clientSourceBasic = [
     'mongo/client/write_concern.cpp',
     'mongo/client/write_operation_base.cpp',
     'mongo/client/write_result.cpp',
-    'mongo/db/jsobj.cpp',
     'mongo/db/json.cpp',
     'mongo/db/dbmessage.cpp',
     'mongo/logger/component_message_log_domain.cpp',
