+++ src/mongo/base/error_codes.err
@@ -96,7 +96,14 @@ error_code("InvalidReplicaSetConfig", 93)
 error_code("NotYetInitialized", 94)
 error_code("NotSecondary", 95)
 error_code("OperationFailed", 96)
-
+error_code("NoProjectionFound", 97)
+error_code("DBPathInUse", 98)
+error_code("WriteConcernNotDefined", 99)
+error_code("CannotSatisfyWriteConcern", 100)
+error_code("OutdatedClient", 101)
+error_code("IncompatibleAuditMetadata", 102)
+error_code("NewReplicaSetConfigurationIncompatible", 103)
+error_code("NodeNotElectable", 104)
 
 # Non-sequential error codes (for compatibility only)
 error_code("NotMaster", 10107) #this comes from assert_util.h
