+++ src/mongo/bson/bsonobj.h
@@ -462,27 +462,6 @@ namespace mongo {
         /** add all elements of the object to the specified list */
         void elems(std::list<BSONElement> &) const;
 
-        /** add all values of the object to the specified vector.  If type mismatches, exception.
-            this is most useful when the BSONObj is an array, but can be used with non-arrays too in theory.
-
-            example:
-              bo sub = y["subobj"].Obj();
-              std::vector<int> myints;
-              sub.Vals(myints);
-        */
-        template <class T>
-        void Vals(std::vector<T> &) const;
-        /** add all values of the object to the specified list.  If type mismatches, exception. */
-        template <class T>
-        void Vals(std::list<T> &) const;
-
-        /** add all values of the object to the specified vector.  If type mismatches, skip. */
-        template <class T>
-        void vals(std::vector<T> &) const;
-        /** add all values of the object to the specified list.  If type mismatches, skip. */
-        template <class T>
-        void vals(std::list<T> &) const;
-
         friend class BSONObjIterator;
         typedef BSONObjIterator iterator;
 
@@ -574,11 +553,11 @@ namespace mongo {
         Status _okForStorage(bool root, bool deep) const;
     };
 
-    std::ostream& operator<<( std::ostream &s, const BSONObj &o );
-    std::ostream& operator<<( std::ostream &s, const BSONElement &e );
+    MONGO_CLIENT_API std::ostream& MONGO_CLIENT_FUNC operator<<( std::ostream &s, const BSONObj &o );
+    MONGO_CLIENT_API std::ostream& MONGO_CLIENT_FUNC operator<<( std::ostream &s, const BSONElement &e );
 
-    StringBuilder& operator<<( StringBuilder &s, const BSONObj &o );
-    StringBuilder& operator<<( StringBuilder &s, const BSONElement &e );
+    MONGO_CLIENT_API StringBuilder& MONGO_CLIENT_FUNC operator<<( StringBuilder &s, const BSONObj &o );
+    MONGO_CLIENT_API StringBuilder& MONGO_CLIENT_FUNC operator<<( StringBuilder &s, const BSONElement &e );
 
 
     struct BSONArray : BSONObj {
