+++ src/mongo/bson/bsonobjiterator.cpp
@@ -1,0 +1,63 @@
+/*    Copyright 2014 10gen Inc.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+#include "mongo/db/jsobj.h"
+
+#include "mongo/util/stringutils.h"
+
+namespace mongo {
+
+    /** Compare two bson elements, provided as const char *'s, by field name. */
+    class BSONIteratorSorted::ElementFieldCmp {
+    public:
+        ElementFieldCmp( bool isArray );
+        bool operator()( const char *s1, const char *s2 ) const;
+    private:
+        LexNumCmp _cmp;
+    };
+    
+    BSONIteratorSorted::ElementFieldCmp::ElementFieldCmp( bool isArray ) :
+    _cmp( !isArray ) {
+    }
+
+    bool BSONIteratorSorted::ElementFieldCmp::operator()( const char *s1, const char *s2 )
+    const {
+        // Skip the type byte and compare field names.
+        return _cmp( s1 + 1, s2 + 1 );
+    }        
+    
+    BSONIteratorSorted::BSONIteratorSorted( const BSONObj &o, const ElementFieldCmp &cmp ) {
+        _nfields = o.nFields();
+        _fields = new const char*[_nfields];
+        int x = 0;
+        BSONObjIterator i( o );
+        while ( i.more() ) {
+            _fields[x++] = i.next().rawdata();
+            verify( _fields[x-1] );
+        }
+        verify( x == _nfields );
+        std::sort( _fields , _fields + _nfields , cmp );
+        _cur = 0;
+    }
+    
+    BSONObjIteratorSorted::BSONObjIteratorSorted( const BSONObj &object ) :
+    BSONIteratorSorted( object, ElementFieldCmp( false ) ) {
+    }
+
+    BSONArrayIteratorSorted::BSONArrayIteratorSorted( const BSONArray &array ) :
+    BSONIteratorSorted( array, ElementFieldCmp( true ) ) {
+    }
+
+} // namespace mongo
