+++ src/mongo/bson/oid.h
@@ -42,6 +42,16 @@ namespace mongo {
     */
     class MONGO_CLIENT_API OID {
     public:
+
+        /**
+         * Functor compatible with std::hash for std::unordered_{map,set}
+         * Warning: The hash function is subject to change. Do not use in cases where hashes need
+         *          to be consistent across versions.
+         */
+        struct Hasher {
+            size_t operator() (const OID& oid) const;
+        };
+
         OID() : a(0), b(0) { }
 
         enum {
