+++ src/mongo/client/examples/insert_demo.cpp
@@ -26,7 +26,6 @@
 
 using namespace std;
 using namespace mongo;
-using namespace bson;
 
 int main(int argc, char* argv[]) {
