+++ src/mongo/db/namespace_string.h
@@ -51,6 +51,23 @@ namespace mongo {
          */
         NamespaceString( const StringData& dbName, const StringData& collectionName );
 
+        /**
+         * Note that these values are derived from the mmap_v1 implementation and that
+         * is the only reason they are constrained as such.
+         */
+        enum MaxNsLenValue {
+            // Maximum possible length of name any namespace, including special ones like $extra.
+            // This includes rum for the NUL byte so it can be used when sizing buffers.
+            MaxNsLenWithNUL = 128,
+
+            // MaxNsLenWithNUL excluding the NUL byte. Use this when comparing std::string lengths.
+            MaxNsLen = MaxNsLenWithNUL - 1,
+
+            // Maximum allowed length of fully qualified namespace name of any real collection.
+            // Does not include NUL so it can be directly compared to std::string lengths.
+            MaxNsCollectionLen = MaxNsLen - 7/*strlen(".$extra")*/,
+        };
+
         StringData db() const;
         StringData coll() const;
