+++ src/mongo/logger/log_component.h
@@ -34,12 +34,12 @@ namespace logger {
             kAccessControl,
             kCommands,
             kIndexing,
-            kJournalling,
             kNetworking,
             kQuery,
             kReplication,
             kSharding,
             kStorage,
+            kJournaling,
             kWrites,
             kNumLogComponents
         };
@@ -49,10 +49,24 @@ namespace logger {
         operator Value() const { return _value; }
 
         /**
+         * Returns parent component.
+         * Returns kNumComponents if parent component is not defined (for kDefault or
+         * kNumLogComponents).
+         */
+        LogComponent parent() const;
+
+        /**
          * Returns short name of log component.
          * Used to generate server parameter names in the format "logLevel_<component short name>".
          */
         std::string getShortName() const;
+
+        /**
+         * Returns dotted name of log component - short name prefixed by dot-separated names of
+         * ancestors.
+         * Used to generate command line and config file option names.
+         */
+        std::string getDottedName() const;
 
     private:
         Value _value;
